improve type safety, prevent access to /welcome if auth payload is not populated
