--- src/components/WalletSelector/mutations/types.ts
@@ -1,6 +1,0 @@
-export type AuthPayloadVariables = {
-  chain: 'Ethereum';
-  address: string;
-  nonce: string;
-  signature: string;
-};
