+++ src/scenes/Auth/Auth.tsx
@@ -50,13 +50,13 @@ function Auth({ queryRef }: Props) {
   // Before the welcome screen, we should preload images so that the animation is smooth
   useEffect(preloadImages, []);
 
-  if (viewer?.__typename === 'Viewer') {
+  if (viewer?.__typename === 'Viewer' && viewer.user) {
     // If user exists in DB, send them to their profile
-    if (viewer.user?.username) {
+    if (viewer.user.username) {
       if (isFeatureEnabled(FeatureFlag.FEED, query)) {
         return <GalleryRedirect to="/home" />;
       }
-      return <GalleryRedirect to={`/${viewer.user?.username}`} />;
+      return <GalleryRedirect to={`/${viewer.user.username}`} />;
     }
 
     // If user is authenticated but hasn't set their username yet.
